requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
idna==3.4
loguru==0.7.2
pyTelegramBotAPI==4.14.0
python-decouple==3.6
requests==2.31.0
telebot==0.0.5
urllib3==2.0.7
win32-setctime==1.1.0
